+++ test/libsolidity/syntaxTests/dataLocations/internalFunction/function_argument_location_specifier_test_internal_transient.sol
@@ -2,4 +2,4 @@ contract test {
     function f(bytes transient) internal {}
 }
 // ----
-// TypeError 6651: (31-46): Data location must be "storage", "memory" or "calldata" for parameter in function, but none was given.
+// TypeError 6651: (31-46): Data location must be "storage", "memory", "calldata" or "constant" for parameter in function, but none was given.