+++ test/libsolidity/syntaxTests/dataLocations/libraries/library_function_with_data_location_transient.sol
@@ -9,10 +9,10 @@ library L {
     function i2() external pure returns (uint[] transient) { }
 }
 // ----
-// TypeError 6651: (28-44): Data location must be "storage", "memory" or "calldata" for parameter in function, but none was given.
-// TypeError 6651: (103-119): Data location must be "storage", "memory" or "calldata" for return parameter in function, but none was given.
-// TypeError 6651: (141-157): Data location must be "storage", "memory" or "calldata" for parameter in function, but none was given.
-// TypeError 6651: (218-234): Data location must be "storage", "memory" or "calldata" for return parameter in function, but none was given.
+// TypeError 6651: (28-44): Data location must be "storage", "memory", "calldata" or "constant" for parameter in function, but none was given.
+// TypeError 6651: (103-119): Data location must be "storage", "memory", "calldata" or "constant" for return parameter in function, but none was given.
+// TypeError 6651: (141-157): Data location must be "storage", "memory", "calldata" or "constant" for parameter in function, but none was given.
+// TypeError 6651: (218-234): Data location must be "storage", "memory", "calldata" or "constant" for return parameter in function, but none was given.
 // TypeError 6651: (256-272): Data location must be "storage", "memory" or "calldata" for parameter in function, but none was given.
 // TypeError 6651: (329-345): Data location must be "storage", "memory" or "calldata" for return parameter in function, but none was given.
 // TypeError 6651: (367-383): Data location must be "storage", "memory" or "calldata" for parameter in external function, but none was given.