+++ test/libsolidity/syntaxTests/dataLocations/libraryInternalFunction/function_argument_location_specifier_test_internal_transient.sol
@@ -2,4 +2,4 @@ library test {
     function f(bytes transient) internal pure {}
 }
 // ----
-// TypeError 6651: (30-45): Data location must be "storage", "memory" or "calldata" for parameter in function, but none was given.
+// TypeError 6651: (30-45): Data location must be "storage", "memory", "calldata" or "constant" for parameter in function, but none was given.