+++ test/libsolidity/syntaxTests/tryCatch/invalid_returns_vars_and_catch_parameter_mutability.sol
@@ -22,9 +22,5 @@ contract C {
     }
 }
 // ----
-// DeclarationError 1788: (90-105): The "constant" keyword can only be used for state variables or variables at file level.
-// DeclarationError 8297: (107-123): The "immutable" keyword can only be used for state variables.
-// DeclarationError 8297: (149-174): The "immutable" keyword can only be used for state variables.
-// DeclarationError 8297: (269-293): The "immutable" keyword can only be used for state variables.
-// DeclarationError 1788: (393-417): The "constant" keyword can only be used for state variables or variables at file level.
-// DeclarationError 1788: (512-535): The "constant" keyword can only be used for state variables or variables at file level.
+// ParserError 3548: (407-415): Location already specified.
+// ParserError 3548: (525-533): Location already specified.